requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
slack-sdk>=3.27.0
requests>=2.31.0
boto3>=1.34.0
pypdf>=4.0.0
